Removing query params
I am trying to simplify the bidi code-base and remove features to make
things easier to debug. There has been a lot of code built up to fit the
particular use-case related to query parameters which I don't believe
should be in the core. It is my view that query parameters should be
orthogonal to resource routing, at least usually.