Allow path chunks to start with a colon
Paths with chunks starting with colons — /like/::this/:one — are valid
but they were not parsed correctly by bidi. The reason was an exception
thrown by the java.net.URI constructor invoked in bidi/just-path.
java.net.URI assumed that a string like ":a" was missing a required URI
schema and threw.  Similar issues occurred in the ClojureScript edition.

This patch solves the problem by prefixing the path with an explicit
schema file:/// before invoking the constructor. Afterwards it drops
the slash which got prepended.

Interestingly enough, replacing bidi/just-path with the following
implementation solves the problem too and according to the test suite
it doesn't appear to be introducing any regressions. This being said,
my gut feeling tells me it might bite us in the future.

   (defn just-path
     [path]
     (apply str (take-while #(not= \? %) path))