Improoved default not-found handler
The default not-found handler will list the roots that the server is
serving.

This leaks some information which may reduce security for
multi-tenanted installations but this can be avoided by providing a
custom not-found handler.

However, as a default in the interests of debugging, it is considered to
be valid trade-off, such that bidi vhosts users can diagnose why their
routes are not being matched.