Accept UUIDs the same way Java does
Some UUID parameters are not matched, most noticeably the "nil" UUID consisting of only zeros. In addition, both Java (and e.g. .NET) and ClojureScript core accept a lot wider variety of "UUIDs", some of which are technically not correct.

This commit proposes to bring the bidi implementation in line with what Java accepts, which is any string with 5 parts separated by hyphens, which each part of a specific length and only hex characters.

Fixes #159